Log-likelihood scores for S-vine copula models

Log-likelihood scores for S-vine copula models

Usage

svinecop_scores(u, model, cores = 1)

Arguments

u

the data; should have approximately uniform margins..

model

model inheriting from class svinecop_dist.

cores

number of cores to use; if larger than one, computations are done in parallel on cores batches .

Value

A matrix containing the score vectors in its rows, where each row corresponds to one observation (row in u). The columns correspond to model parameters in the order: copula parameters of first tree, copula parameters of second tree, etc. Duplicated parameters in the copula model are omitted.

See also

Examples

# load data set
data(returns)  

# convert to uniform margins
u <- pseudo_obs(returns[1:100, 1:3])

# fit parametric S-vine copula model with Markov order 1
fit <- svinecop(u, p = 1, family_set = "parametric")

svinecop_loglik(u, fit)
#> [1] 91.58517
svinecop_scores(u, fit)
#>                [,1]         [,2]         [,3]         [,4]        [,5]
#>   [1,]  1.324513677 -1.177262570  0.159468681  0.708186149  1.40306548
#>   [2,] -0.478180704 -0.315073706 -0.283158197  0.948076232 -1.26166320
#>   [3,] -1.938628197 -0.299742996 -0.341622965 -0.594563560 -1.80011224
#>   [4,]  0.543175120 -1.315666200 -0.348152310 -1.591056588  0.58738874
#>   [5,]  0.920405777 -0.981873167 -0.028613151  1.657957569  1.20474065
#>   [6,] -1.288208922  0.069224680 -0.176641328  0.904083650 -1.03062156
#>   [7,]  0.736978025 -1.357634897 -0.256673848 -2.029995796  1.24683914
#>   [8,] -1.124436211  0.387107992  0.071023430  0.960566314 -0.30648176
#>   [9,] -1.157792618  0.242416043 -0.098506884 -0.342285728 -0.97625605
#>  [10,] -0.750483813  0.798103200  0.257362951 -0.840707030  0.07386860
#>  [11,] -1.075208352  0.450103161 -0.102245730 -0.904729723  1.17204637
#>  [12,] -0.001285325 -0.593248134 -0.277818241 -0.060337319  1.28174803
#>  [13,] -1.080447431  0.289452003 -0.087424364 -0.885149065 -0.35325937
#>  [14,]  1.611108979  0.295599256  0.647741905 -0.411058223 -0.39767023
#>  [15,]  0.635196774 -0.905817903 -0.136896351  0.345201841 -1.10222810
#>  [16,]  1.679208418  0.565809822  0.543735149 -0.554961387  1.27819091
#>  [17,]  0.278277206 -0.752985140 -0.229720765  1.390820757 -0.96477336
#>  [18,] -0.919852042  0.637681786  0.095050957 -0.881636902  1.08921928
#>  [19,] -0.943899006 -0.082165301 -0.438456052  0.859164880  1.42321983
#>  [20,]  1.790619121  2.065914719  1.020342507  1.067031369 -1.70168295
#>  [21,] -1.009838823  0.273394112 -0.122118713 -0.336528163 -0.65185019
#>  [22,] -1.041709385  0.174527819 -0.309552346 -0.078146702 -0.78733623
#>  [23,]  0.874417377 -1.324609925 -0.169181958 -1.867271857  1.19881568
#>  [24,] -1.065847924  0.531189783 -0.036457016  1.324703255 -0.51152272
#>  [25,]  0.826023952 -1.482928314 -0.239337052  0.244275147 -1.32008829
#>  [26,]  3.006611325  0.278014948  0.340219064 -0.332803345 -0.93937067
#>  [27,]  0.975669438 -1.142663886 -0.057097911  0.740252332 -1.57303948
#>  [28,]  0.220085964 -0.769632751 -0.214634141 -1.213484545  0.78809493
#>  [29,]  0.356559113 -0.831972964 -0.256757810  1.093260243  1.28842873
#>  [30,]  0.287739754 -1.017493075 -0.374879518 -1.879632478  1.12384393
#>  [31,] -1.031388178  0.289197012 -0.212961137  0.233723146 -1.00204659
#>  [32,]  2.613588329  0.344162583  0.382027715  1.102044806 -0.55204623
#>  [33,] -0.829782413  0.656023635  0.137571141 -0.545335902 -0.37311700
#>  [34,] -0.115780268 -0.572008389 -0.360745708  1.555310818 -1.07098085
#>  [35,] -0.148769929 -0.530853540 -0.277914724 -0.715536328 -0.09812347
#>  [36,] -0.011086005 -0.608017135 -0.266072201 -1.011732158  0.30347393
#>  [37,]  0.657452100  0.530396689  0.366638945  0.558088439  4.47954238
#>  [38,]  3.521231124  0.193326847  0.271879685  0.903449820  2.20834056
#>  [39,] -1.030213944  0.436932757 -0.076343761  0.353816361 -1.07271836
#>  [40,]  1.818190471  2.003673203  1.057859493 -1.561736493  0.98521898
#>  [41,] -1.734895359  0.794175978 -0.048509740 -1.670963939  1.20635566
#>  [42,] -0.948316866  0.086325304 -0.195327222  0.944734503 -1.15718004
#>  [43,] -0.845651085  0.762718160  0.200489493  0.384318802 -1.09066400
#>  [44,] -0.886176672  0.152444441 -0.229967805  0.407285389  1.24376732
#>  [45,] -0.378249417 -0.330706020 -0.335304630  0.456277068  1.32231649
#>  [46,]  1.619734075 -1.479325062  0.408721772 -2.114424672  1.34439960
#>  [47,] -0.363537927 -0.747860970 -0.304565553 -1.266675658  0.15940539
#>  [48,] -0.956017727  0.541590893  0.036386190  0.002964137  8.12609601
#>  [49,] -1.008919516  0.320606121 -0.105002435  1.104645322  1.25934399
#>  [50,]  1.014442317 -0.918187531  0.196304735  0.257760338 -1.13098819
#>  [51,] -0.297345293 -0.543530147 -0.431234169  1.020686708 -1.91615092
#>  [52,] -0.985166361  0.098452081 -0.339636885  0.152035636  1.42476399
#>  [53,]  0.547627500 -0.916288920 -0.197410516 -1.116297070  1.33179693
#>  [54,] -1.216029960 -0.086839205 -0.239818410  0.952659074 -0.93675716
#>  [55,]  0.627224173 -1.270723139 -0.290804981  0.970212731  1.42260300
#>  [56,] -0.068246784 -0.623515804 -0.263423102 -1.197313138  0.86636231
#>  [57,] -0.685705330 -0.147255134 -0.373545176 -0.715786517 -0.13117632
#>  [58,] -0.062007608 -0.736106701 -0.419732654 -0.845005908 -0.10142685
#>  [59,] -0.245883556  0.720035536  0.322471319  1.147626657 -0.95790111
#>  [60,]  0.305356504 -0.816982894 -0.191397663 -0.581360819 -0.26728725
#>  [61,]  1.132492245 -1.353348602 -0.015419305 -0.815122125 -0.12429621
#>  [62,] -1.038730710  0.258525829 -0.117210986  0.541427767 -1.21813290
#>  [63,] -1.124851010  0.543959677 -0.057537577  0.263708723 -1.00830491
#>  [64,] -0.196652337 -0.456654531 -0.304676757 -0.857305991  0.11268463
#>  [65,] -0.193425599 -0.458638686 -0.306742536 -0.861928103  0.11821659
#>  [66,] -0.189903462 -0.461164905 -0.308856493 -0.867108833  0.12403618
#>  [67,]  0.029810775 -0.658904068 -0.335519825  1.549418498 -1.27454741
#>  [68,] -0.990722636  0.307210303 -0.130352153 -0.045261563 -0.80187845
#>  [69,]  1.406712379 -0.313742523  0.426172210  0.773909672 -1.32442017
#>  [70,] -0.155682281 -0.527657725 -0.353617785 -0.562299098 -0.27742080
#>  [71,] -0.917344100 -0.002924229 -0.379767087 -0.355997303 -0.55618782
#>  [72,] -1.135978937  0.650459952  0.018219388 -0.906004229  0.17273987
#>  [73,] -1.827871852  1.052536765  0.255535598  0.835787973 -1.72163529
#>  [74,] -2.016568626  0.675345232 -0.244052509 -1.793505563  0.79097719
#>  [75,] -0.645951056 -0.363159705 -0.289303441  0.329726509 -1.13897515
#>  [76,]  0.812639493 -1.224782453 -0.175704137  1.460921463 -2.17147412
#>  [77,] -1.459644712  0.811187507  0.054079468  0.729235350 -1.47552047
#>  [78,] -1.190483477  0.094870004 -0.418007679 -1.466289555  1.67118673
#>  [79,] -0.330674944  0.950981900  0.722811396 -0.922402501  0.06645370
#>  [80,] -1.111502957  0.086514507 -0.178793116 -1.352308369  0.10907218
#>  [81,]  1.099034005 -1.749452380 -0.123849425 -2.518415621  1.76887396
#>  [82,] -1.737902455  1.112069485  0.412388246  0.891642671 -1.78145421
#>  [83,] -0.146427726  0.654267630  0.308185973 -1.357013255  0.40107682
#>  [84,]  0.987772840 -0.943989240  0.023950876 -0.319777149 -0.49373699
#>  [85,]  0.995007017 -0.950400577  0.024453277 -0.415133690 -0.39760633
#>  [86,] -2.249077260  1.294191142  0.748257215  1.647671117 -2.59119519
#>  [87,]  1.157760735 -0.693170582  0.261936722 -0.220643117 -0.68391528
#>  [88,]  4.818630344  0.000833635  0.003140519 -0.844198980 -1.00842504
#>  [89,]  1.127230882 -0.737027268  0.228062824  0.262965837 -1.02996939
#>  [90,]  1.933688265  3.146807356  2.031194182  8.961782532  0.21234174
#>  [91,]  1.528139232  0.620774917  0.698742706  0.558277816 -1.40890901
#>  [92,]  1.204299420  0.569695076  0.446530495  1.134214730 -0.78735228
#>  [93,]  0.635859302 -1.020229759 -0.048879913  0.669218798  1.07986042
#>  [94,] -1.125051164  0.935023982  0.306754531 -0.003877932 -0.82000650
#>  [95,]  1.547279059 -0.174351792  0.522364123  0.427774094  5.74496658
#>  [96,] -1.511532071 -0.005322960 -0.206253874 -0.927980971 -1.09805278
#>  [97,] -0.533335673  0.781379627  0.300615114  0.647749150 -1.23649197
#>  [98,]  0.236983769 -0.733224961 -0.251931418  0.927719179 -1.48057817
#>  [99,] -1.093759926  0.283788746 -0.250238435 -1.369761684  0.89962158
#>                [,6]         [,7]         [,8]        [,9]        [,10]
#>   [1,]  0.634437023 -0.309469402 -0.200468807  0.03013514 -0.572131684
#>   [2,]  0.034321436  0.232360235  0.069625969  1.86210160 -0.562350185
#>   [3,] -0.341777506 -0.905300979 -0.138031565 -0.52599279 -0.424589303
#>   [4,] -0.302784462  2.154558976 -0.042198496 -0.33990726 -0.047331537
#>   [5,]  0.584224999  1.209412557 -0.198543977  0.49517507 -0.069892587
#>   [6,]  0.150522516 -1.108936615  0.032336624  1.26631461 -0.399960567
#>   [7,] -0.036212966  0.722070441 -0.063041330 -0.37371636 -0.010386563
#>   [8,]  0.314319452  0.800109212 -0.008843908  0.70733030  0.017979473
#>   [9,] -0.260511409 -1.584175996 -0.042389424 -0.23693280  0.131876262
#>  [10,] -0.239010816 -0.816085385  0.112006229  0.79699369 -0.081947691
#>  [11,]  0.210963622 -0.563067327  0.186664488 -0.10775803 -1.068299391
#>  [12,]  0.410592654 -1.700259796 -0.075664699  0.19409075  0.188462196
#>  [13,] -0.260265328 -2.258645444 -0.008168656  0.17256536  0.132784218
#>  [14,] -0.306139319 -0.022144550  0.102048043 -0.43586957 -0.008722228
#>  [15,] -0.236813710  0.186971829  0.105563700 -0.15551997 -1.649081696
#>  [16,]  0.324338418  0.312850852  0.465794498  2.19447001  0.999877779
#>  [17,]  0.377159114  0.658853754  0.228642778  0.74878891  1.094982919
#>  [18,]  0.206565812  0.739028525  0.258978157  0.60835939  0.872127950
#>  [19,]  0.725557226 -0.134551168  0.203500406 -0.45125190 -0.865669355
#>  [20,] -0.035418488  0.006993861  0.410038578  2.80221502  0.823943237
#>  [21,] -0.255687875  0.648205545 -0.031532664 -0.18032480 -0.098618061
#>  [22,] -0.331033540  0.524623031 -0.072553195 -0.05581031 -0.006217771
#>  [23,] -0.031901002 -0.443957063 -0.076304300 -0.20377118  0.258019569
#>  [24,]  0.413862764 -0.379302102 -0.113378596  0.12562713  0.179325839
#>  [25,] -0.369392974 -1.527014171  0.072011421 -0.50897409  0.245674523
#>  [26,] -0.258286422  1.465971485  0.172574576  0.80891848  0.594399226
#>  [27,] -0.193458074  0.910689926  0.176491685 -0.52442794  1.072384371
#>  [28,] -0.022988668  0.783963175 -0.148891461 -0.34216370 -0.623201024
#>  [29,]  0.573519777 -1.040630677 -0.071165914 -0.05486855 -0.021694035
#>  [30,] -0.082799735  0.702022204 -0.041356621  0.88725962  0.096361973
#>  [31,] -0.251750181 -0.166403719 -0.070406626  0.02979369  0.112084419
#>  [32,]  0.288209032  0.289598043 -0.191002928  0.72745537 -0.631397185
#>  [33,] -0.259681617 -0.129148946  0.045382334  0.45832339 -1.283751737
#>  [34,]  0.597537618  0.727748495  0.087415086 -0.29119480  0.338847279
#>  [35,] -0.253334377  0.477228414 -0.061491616 -0.45110371 -0.089412105
#>  [36,] -0.194986954 -0.186042134 -0.081949976 -0.20643166 -0.133052415
#>  [37,]  0.432282140  0.323507639 -0.169398181 -0.20198329  0.324191553
#>  [38,]  0.507658868  0.212060613 -0.280034139  1.52502669 -0.953317874
#>  [39,] -0.182867655  0.253569296 -0.061917745  0.02724968 -0.747393911
#>  [40,] -0.084364398  0.291966175 -0.094188840 -0.51555588  0.206513947
#>  [41,]  0.019437309  0.065799134  0.193677154 -1.12310411  1.553246238
#>  [42,]  0.080426954  0.308202935  0.008842702  0.59511895  0.376423874
#>  [43,] -0.189979265  0.212060613 -0.082773969  0.52378602  0.292607996
#>  [44,]  0.461014181 -0.331033841  0.142275973  0.03815342 -0.624432184
#>  [45,]  0.506036424 -1.851216971  0.051218181  0.06164249 -0.062013854
#>  [46,]  0.015768253  0.847477770 -0.103619843  1.27833092  0.303855349
#>  [47,] -0.180687007  1.209412557 -0.181918775  1.17875745 -1.030626524
#>  [48,]  0.006730852  2.314868671 -0.079522273 -0.20581669  0.175915902
#>  [49,]  0.538650276 -2.404552076 -0.001443063 -0.32013572  0.127440109
#>  [50,] -0.167562200  0.425031971  0.112512527 -0.84400090 -0.640741821
#>  [51,] -0.119499301  0.754631127 -0.019409327 -0.28982202 -0.486162123
#>  [52,]  0.541498130  0.382338286  0.191991413  0.17564912  0.104556807
#>  [53,]  0.231316543  0.018709301 -0.108319649  1.31468013  0.432353280
#>  [54,]  0.179078750 -0.156558601  0.069349954  0.82278206 -0.241253819
#>  [55,]  0.776740997  0.306867211 -0.038778382 -0.23300933  0.121381217
#>  [56,]  0.027889730  0.228545126 -0.131348527  2.58841103  0.063393579
#>  [57,] -0.334369496  0.490512310 -0.043618040 -0.17888918  0.222964492
#>  [58,] -0.375484537 -0.597707966 -0.042203787 -0.13447697  0.106819196
#>  [59,]  0.213107752 -0.008677463 -0.166966430 -0.33035044  0.022459546
#>  [60,] -0.262241425  0.190653183 -0.048927305 -0.60140698  0.680146490
#>  [61,] -0.374017740  0.508738554  0.025722472 -0.25342568 -0.738884184
#>  [62,] -0.090416461  0.555068515 -0.037557594 -0.07023516 -0.018152776
#>  [63,] -0.227966202  0.191476239 -0.074530381  0.13872246  0.104079814
#>  [64,] -0.251831537  0.249538306 -0.051006885 -0.28928412  0.170235493
#>  [65,] -0.253227990  0.198703867 -0.050265849 -0.13331549  0.073601660
#>  [66,] -0.254609462  0.212699228 -0.049581155 -0.12988324  0.078247992
#>  [67,]  0.510320644  0.066853255  0.118700488 -0.07129967 -0.012824962
#>  [68,] -0.240756840  0.455057476 -0.041610329 -0.11058019  0.242523504
#>  [69,] -0.085652594  0.276620308  0.297154566 -0.45006114  0.014765768
#>  [70,] -0.332489861  0.126756214 -0.037841354 -0.28828516 -0.397617758
#>  [71,] -0.366381972 -0.287777227 -0.064822496 -0.20025518  0.144127210
#>  [72,] -0.233451962  0.355435259  0.067530491 -0.12153278 -0.402451184
#>  [73,] -0.181119094  0.316824957 -0.216296523  0.20748817 -1.353804623
#>  [74,] -0.252175581 -2.718863600  0.063863082  0.58550039 -1.803585754
#>  [75,] -0.154628717  0.649654304  0.007023435  0.64308267  0.749888661
#>  [76,]  0.161611077  0.657473413  0.202796114 -0.51164818 -0.520845810
#>  [77,] -0.169458646 -1.383113814 -0.160917461  0.47518437  0.784689651
#>  [78,]  0.494966463 -0.348827375  0.189845083  0.28697682 -1.042794353
#>  [79,] -0.328895631 -0.290903192  0.059856282 -5.77015539 -1.629556049
#>  [80,] -0.189198556  0.590403907 -0.037867406 -0.08632957 -0.392256810
#>  [81,]  0.368408074  1.756609085 -0.093969493 -0.28156722  0.112368361
#>  [82,] -0.164388607 -2.512416224 -0.232898224 -3.86092462  0.333622863
#>  [83,] -0.114349719  1.357056269  0.257012331  0.42177419 -1.792058282
#>  [84,] -0.308582204 -0.103306439  0.054998761 -0.18960657  1.016826502
#>  [85,] -0.310784274  0.205883966  0.042231149 -0.16984454 -1.225799800
#>  [86,]  0.276141716 -1.063011702 -0.353866656  1.50518543  1.270397841
#>  [87,] -0.252496557  0.534080754  0.037978772 -0.60063131  2.350615180
#>  [88,] -0.322397204 -1.891070226  0.363415730  4.98129754  2.057808332
#>  [89,] -0.191197614 -1.369974010  0.148497507 -0.66518823  2.063641215
#>  [90,]  0.294543980  0.383313887 -0.332013970 -0.53563987 -2.979658280
#>  [91,] -0.241549937 -0.626457640 -0.184456449 -7.65315370  1.026465413
#>  [92,]  0.248186350  0.179552524 -0.190586881  0.44003275 -2.247798569
#>  [93,]  0.450412081 -1.505034680 -0.336630549 -0.71403582  1.220952135
#>  [94,] -0.303858216 -0.212128294 -0.071809493 -0.83177139  1.378766398
#>  [95,]  0.447899030 -1.369974010  0.517868440 -0.22161339  1.672191928
#>  [96,] -0.340242192  2.409437745 -0.099840108 -0.42807083 -0.723863627
#>  [97,] -0.110050585 -0.777534730 -0.113170612  0.60628186  0.341065266
#>  [98,] -0.056883766  0.216819889  0.120432058 -0.45711981  0.650644796
#>  [99,] -0.070266782  0.241647205  0.079779604  0.18261232  0.584660267
#>                [,11]
#>   [1,]  0.0020229183
#>   [2,]  0.0021095805
#>   [3,]  0.0032358241
#>   [4,]  0.0049210833
#>   [5,]  0.0065874901
#>   [6,]  0.0038494436
#>   [7,]  0.0089669632
#>   [8,]  0.0106719282
#>   [9,]  0.0124242159
#>  [10,] -0.0083141817
#>  [11,] -0.0011137335
#>  [12,]  0.0056369405
#>  [13,]  0.0126882806
#>  [14,] -0.0347088094
#>  [15,] -0.0073452473
#>  [16,] -0.0297265339
#>  [17,] -0.0212296233
#>  [18,]  0.0048226511
#>  [19,]  0.0001024668
#>  [20,] -0.0430771756
#>  [21,] -0.0574242845
#>  [22,]  0.0091542668
#>  [23,]  0.0103588348
#>  [24,]  0.0112534481
#>  [25,]  0.0099596054
#>  [26,] -0.0186975855
#>  [27,] -0.0044006827
#>  [28,]  0.0002132837
#>  [29,]  0.0073857870
#>  [30,]  0.0123653193
#>  [31,]  0.0121575360
#>  [32,] -0.0128877576
#>  [33,] -0.0012384282
#>  [34,]  0.0062103998
#>  [35,]  0.0086542448
#>  [36,]  0.0078651553
#>  [37,]  0.0092098509
#>  [38,] -0.0144101453
#>  [39,] -0.0130448142
#>  [40,] -0.0334796837
#>  [41,]  0.0060492181
#>  [42,] -0.0064049714
#>  [43,]  0.0018569891
#>  [44,]  0.0007467263
#>  [45,]  0.0078928386
#>  [46,] -0.0011635078
#>  [47,]  0.0002925344
#>  [48,]  0.0048276304
#>  [49,]  0.0127357334
#>  [50,] -0.0549635190
#>  [51,] -0.0584742328
#>  [52,]  0.0039356789
#>  [53,]  0.0063976089
#>  [54,]  0.0044448215
#>  [55,]  0.0076703221
#>  [56,]  0.0072483172
#>  [57,]  0.0106127543
#>  [58,]  0.0119113029
#>  [59,]  0.0044287380
#>  [60,]  0.0057696669
#>  [61,]  0.0006942949
#>  [62,]  0.0049119854
#>  [63,]  0.0079547773
#>  [64,]  0.0093429344
#>  [65,]  0.0118904802
#>  [66,]  0.0119898772
#>  [67,]  0.0090905711
#>  [68,]  0.0100437860
#>  [69,] -0.0372815633
#>  [70,] -0.0387068983
#>  [71,]  0.0109193484
#>  [72,]  0.0008706757
#>  [73,] -0.0017468202
#>  [74,]  0.0112325638
#>  [75,]  0.0004984222
#>  [76,]  0.0023481495
#>  [77,]  0.0042984253
#>  [78,]  0.0004894474
#>  [79,] -0.0042623680
#>  [80,] -0.0494832080
#>  [81,]  0.0118049769
#>  [82,] -0.0045120033
#>  [83,]  0.0081252678
#>  [84,]  0.0057863716
#>  [85,]  0.0014375836
#>  [86,] -0.0138548965
#>  [87,]  0.0548148702
#>  [88,]  0.0190542098
#>  [89,]  0.0203926514
#>  [90,]  0.0792720660
#>  [91,] -0.0237122740
#>  [92,]  0.0105996027
#>  [93,]  0.0074099991
#>  [94,]  0.0102693156
#>  [95,]  0.0127461376
#>  [96,] -0.0325271256
#>  [97,] -0.0002018375
#>  [98,]  0.0041230438
#>  [99,]  0.0063339351
svinecop_hessian(u, fit)
#>               [,1]         [,2]          [,3]          [,4]          [,5]
#>  [1,] -1.508242340  0.245722797 -0.2123586991  0.0000000000  0.0000000000
#>  [2,]  0.245722797 -0.358543760 -0.1894417200  0.0000000000  0.0000000000
#>  [3,] -0.212358699 -0.189441720 -0.1439333529  0.0000000000  0.0000000000
#>  [4,]  0.000000000  0.000000000  0.0000000000 -1.8370278721  0.4395014319
#>  [5,]  0.000000000  0.000000000  0.0000000000  0.4395014319 -2.3392976513
#>  [6,]  0.000000000  0.000000000  0.0000000000 -0.2396947012 -0.1228376978
#>  [7,]  0.000000000  0.000000000  0.0000000000  0.0000000000  0.0000000000
#>  [8,]  0.011211692 -0.005040302 -0.0023481941 -0.0097926631 -0.0083971654
#>  [9,] -0.139804302 -0.021389630 -0.0121718685  0.0083448944  0.0011731269
#> [10,]  0.080825493  0.174042095  0.0286060187  0.0063905512 -0.0119393740
#> [11,]  0.002723608 -0.007959599  0.0004965833 -0.0005300579 -0.0001169995
#>                [,6]      [,7]          [,8]         [,9]        [,10]
#>  [1,]  0.0000000000  0.000000  0.0000000000  0.000000000  0.000000000
#>  [2,]  0.0000000000  0.000000  0.0000000000  0.000000000  0.000000000
#>  [3,]  0.0000000000  0.000000  0.0000000000  0.000000000  0.000000000
#>  [4,] -0.2396947012  0.000000  0.0000000000  0.000000000  0.000000000
#>  [5,] -0.1228376978  0.000000  0.0000000000  0.000000000  0.000000000
#>  [6,] -0.0934714361  0.000000  0.0000000000  0.000000000  0.000000000
#>  [7,]  0.0000000000 -1.006142  0.0000000000  0.000000000  0.000000000
#>  [8,] -0.0046244418  0.000000 -0.0253267513  0.000000000  0.000000000
#>  [9,]  0.0002681758  0.000000 -0.0170659561 -1.970676783  0.000000000
#> [10,] -0.0008669042  0.000000  0.0078621156  0.171051787 -0.777496857
#> [11,] -0.0001959292  0.000000 -0.0002101492  0.009046062 -0.007246873
#>               [,11]
#>  [1,]  0.0000000000
#>  [2,]  0.0000000000
#>  [3,]  0.0000000000
#>  [4,]  0.0000000000
#>  [5,]  0.0000000000
#>  [6,]  0.0000000000
#>  [7,]  0.0000000000
#>  [8,]  0.0000000000
#>  [9,]  0.0000000000
#> [10,] -0.0072468726
#> [11,] -0.0005493726